Stage 1 (Part)
Part of stage 1 completed. Good enough for a trial.

Deleters (Our  "Bottomless Pits") and Firewalls ("Our main stationary
obstacles) are lack script to cause death to the player

Scrip to stick player to moving platforms added. It's a little buggy
when teleporting too close to the platforms edge, messing the shape of
the character until the next blink  (But I was convinced to consider it
a feature, not a bug. But seriously, I'll look into it later)